refactor plugin support into its own module
